Move the install package into workflow/startup
